More preexisting installation fixes, and documentation!

+++ bash/install/main.sh
@@ -1,10 +1,10 @@
 #!/bin/bash -e
 
 if [ "$EUID" -ne 0 ]; then
-	echo Installing Starship
+	echo Installing Starship and coreutils
 
 	if brew commands &> /dev/null; then
-		brew install starship
+		brew install starship coreutils
 	else
 		sh -c "$(curl -fsSL https://starship.rs/install.sh)"
 	fi
@@ -16,41 +16,44 @@ fi
 
 echo Archiving old Bash configuration...
 
-if [ ! -d "${HOME}"/.shell-config-archive ]; then
+# Check if we're archiving an old installation directory first
+if [ -d "${HOME}/.bash-config" ]; then
+	mv -v "${HOME}/.bash-config" "${HOME}/.shell-config-archive"
+elif [ ! -d "${HOME}"/.shell-config-archive ]; then
 	mkdir -v "${HOME}"/.shell-config-archive
 fi
 
-if [ -f "${HOME}"/.bash_aliases ]; then
+if [ -f "${HOME}"/.bash_aliases ] && [ ! -L "${HOME}"/.bash_aliases ]; then
 	mv -v "${HOME}"/.bash_aliases "${HOME}"/.shell-config-archive/bash_aliases
 fi
 
-if [ -f "${HOME}"/.bash_exports ]; then
+if [ -f "${HOME}"/.bash_exports ] && [ ! -L "${HOME}"/.bash_exports ]; then
 	mv -v "${HOME}"/.bash_exports "${HOME}"/.shell-config-archive/bash_exports
 fi
 
-if [ -f "${HOME}"/.bash_logout ]; then
+if [ -f "${HOME}"/.bash_logout ] && [ ! -L "${HOME}"/.bash_logout ]; then
 	mv -v "${HOME}"/.bash_logout "${HOME}"/.shell-config-archive/bash_logout
 fi
 
-if [ -f "${HOME}"/.bash_profile ]; then
+if [ -f "${HOME}"/.bash_profile ] && [ ! -L "${HOME}"/.bash_profile ]; then
 	mv -v "${HOME}"/.bash_profile "${HOME}"/.shell-config-archive/bash_profile
 fi
 
-if [ -f "${HOME}"/.bashrc ]; then
+if [ -f "${HOME}"/.bashrc ] && [ ! -L "${HOME}"/.bashrc ]; then
 	mv -v "${HOME}"/.bashrc "${HOME}"/.shell-config-archive/bashrc
 fi
 
-if [ -f "${HOME}"/.config/starship.toml ]; then
+if [ -f "${HOME}"/.config/starship.toml ] && [ ! -L "${HOME}"/.config/starship.toml ]; then
 	mv -v "${HOME}"/.config/starship.toml "${HOME}"/.shell-config-archive/starship.toml
 fi
 
-if [ -f "${HOME}"/.vimrc ]; then
+if [ -f "${HOME}"/.vimrc ] && [ ! -L "${HOME}"/.vimrc ]; then
 	mv -v "${HOME}"/.vimrc "${HOME}"/.shell-config-archive/vimrc
 fi
 
 echo Checking and symlinking repository
 
-REPOSITORY_DIRECTORY="$(dirname "$(dirname "$(dirname "${0}")")")"
+REPOSITORY_DIRECTORY="$(dirname "$(dirname "$(dirname "$(realpath "${0}")")")")"
 
 if [ "${REPOSITORY_DIRECTORY}" != "${HOME}"/.shell-config ]; then
 	ln -snfv "${REPOSITORY_DIRECTORY}" "${HOME}"/.shell-config
